md5_plain: resolve casting warning for uint64_t to uint8_t

+++ hash_functions/md5_plain.c
@@ -26,14 +26,14 @@
 #endif
 
 #define PUT_64BIT_LE(cp, value) do {					\
-	(cp)[7] = (value) >> 56;					\
-	(cp)[6] = (value) >> 48;					\
-	(cp)[5] = (value) >> 40;					\
-	(cp)[4] = (value) >> 32;					\
-	(cp)[3] = (value) >> 24;					\
-	(cp)[2] = (value) >> 16;					\
-	(cp)[1] = (value) >> 8;						\
-	(cp)[0] = (value); } while (0)
+	(cp)[7] = (uint8_t)((value) >> 56);					\
+	(cp)[6] = (uint8_t)((value) >> 48);					\
+	(cp)[5] = (uint8_t)((value) >> 40);					\
+	(cp)[4] = (uint8_t)((value) >> 32);					\
+	(cp)[3] = (uint8_t)((value) >> 24);					\
+	(cp)[2] = (uint8_t)((value) >> 16);					\
+	(cp)[1] = (uint8_t)((value) >> 8);						\
+	(cp)[0] = (uint8_t)((value)); } while (0)
 
 #define PUT_32BIT_LE(cp, value) do {					\
 	(cp)[3] = (value) >> 24;					\
